+++ Gotchi/Portfolios/Mangers/IPortfolioManger.cs
@@ -11,7 +11,7 @@ namespace Gotchi.Portfolios.Managers
         Portfolio GetByPortfolioId(string? portfolioId);
         IEnumerable<Portfolio> GetByPersonId(string personId);
         void RemovePortfolioAsset(Portfolio portfolio, string coinMarketId);
-        void SellAsset(Portfolio portfolio, CryptoCoin coin, int units);
+        void SellAsset(Portfolio portfolio, CryptoCoin coin, float units);
         bool Store(Portfolio portfolio);
         void Update(Portfolio portfolio);
         float WithdrawFromAccount(Portfolio portfolio, float amount);
